+++ src/main/java/store/onuljang/controller/response/AdminProductListItems.java
@@ -3,6 +3,7 @@ package store.onuljang.controller.response;
 import lombok.Builder;
 import store.onuljang.repository.entity.Product;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -15,7 +16,7 @@ public record AdminProductListItems(
     public record AdminProductListItem(
         long id,
         String name,
-        long price,
+        BigDecimal price,
         int stock,
         long totalSold,
         String productUrl,
